requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.0.300
langchain-community==0.0.38
langchain-huggingface==0.0.3
langchain-openai==0.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
openai